spnfs: fix /proc _read and _write function return types
Return type should be ssize_t, not int.  Fixes spnfsd restart bug
on 64 bit machines.

+++ fs/nfsd/spnfs_com.c
@@ -314,8 +314,8 @@ int spnfs_enabled(void)
  */
 
 /*************** start ctl **************************/
-static int ctl_write(struct file *file, const char __user *buf, size_t count,
-		       loff_t *offset)
+static ssize_t ctl_write(struct file *file, const char __user *buf,
+			 size_t count, loff_t *offset)
 {
 	int cmd, rc;
 
@@ -347,8 +347,8 @@ static int getfh_open(struct inode *inode, struct file *file)
 	return 0;
 }
 
-static int getfh_read(struct file *file, char __user *buf, size_t count,
-		      loff_t *offset)
+static ssize_t getfh_read(struct file *file, char __user *buf, size_t count,
+			  loff_t *offset)
 {
 	if (copy_to_user(buf, file->private_data, sizeof(struct nfs_fh)))
 		return -EFAULT;
@@ -356,8 +356,8 @@ static int getfh_read(struct file *file, char __user *buf, size_t count,
 	return count;
 }
 
-static int getfh_write(struct file *file, const char __user *buf, size_t count,
-		       loff_t *offset)
+static ssize_t getfh_write(struct file *file, const char __user *buf,
+			   size_t count, loff_t *offset)
 {
 	int fd;
 
@@ -385,8 +385,8 @@ static struct file_operations getfh_ops = {
 
 
 /*************** start recall layout ***************/
-static int recall_write(struct file *file, const char __user *buf, size_t count,
-			loff_t *offset)
+static ssize_t recall_write(struct file *file, const char __user *buf,
+			    size_t count, loff_t *offset)
 {
 	char path[128];
 	char *p;
@@ -419,8 +419,8 @@ static struct file_operations recall_ops = {
 
 #ifdef CONFIG_SPNFS_LAYOUTSEGMENTS
 /*************** start layoutseg **************************/
-static int layoutseg_write(struct file *file, const char __user *buf,
-				size_t count, loff_t *offset)
+static ssize_t layoutseg_write(struct file *file, const char __user *buf,
+			       size_t count, loff_t *offset)
 {
 	char cmd[3];
 
@@ -440,8 +440,8 @@ static struct file_operations layoutseg_ops = {
 /*************** end layoutseg ****************************/
 
 /*************** start layoutsegsize **************************/
-static int layoutsegsize_write(struct file *file, const char __user *buf,
-				size_t count, loff_t *offset)
+static ssize_t layoutsegsize_write(struct file *file, const char __user *buf,
+				   size_t count, loff_t *offset)
 {
 	char cmd[50];
